+++ CHANGELOG.md
@@ -2,6 +2,18 @@
 
 This format is based on [Keep A Changelog](https://keepachangelog.com/en/1.0.0).
 
+## 1.0.1 -- 2022-05-17
+
+### Added
+
+* `alwaysFailProperty` -- an universial property that ensures script always fails
+
+### Fixed
+
+* Bug that made `classifiedProperty` fail to differentiate between
+  wrong result and expected crash when script ran without
+  crashing. (Fixed #5)
+
 ## 1.0.0 -- 2022-05-05
 
 ### Added